Fixed Left Bar not collapsing
The click function was being registered 2 times for each heading causing
it to collapse then expand again. This fix remove the first registration
on second call. Real fix is not to call scripts.php 2 times in one page
but it is in both the header.php and the page.php